refactor: separate jsondb function into historydata.go
