+++ internal/persistence/jsondb/setup_test.go
@@ -54,7 +54,8 @@ type dagTestHelper struct {
 func (d dagTestHelper) Writer(t *testing.T, requestID string, startedAt time.Time) writerTestHelper {
 	t.Helper()
 
-	filePath := d.th.DB.generateFilePath(d.th.Context, d.DAG.Location, newUTC(startedAt), requestID)
+	data := d.th.DB.Data(d.th.Context, d.DAG.Location)
+	filePath := data.generateFilePath(d.th.Context, newUTC(startedAt), requestID)
 	writer := NewWriter(filePath)
 	require.NoError(t, writer.Open())
